# Provide contextualized revision data on the request's context.