--- a/cmd/activator/main.go
+++ b/cmd/activator/main.go
@@ -244,17 +244,16 @@ func main() {
 	ah = reqLogHandler
 	ah = &activatorhandler.ProbeHandler{NextHandler: ah}
 
-	sigCtx, sigCancel := context.WithCancel(context.Background())
+	// NOTE: MetricHandler is being used as the outermost handler of the meaty bits. We're not interested in measuring
+	// the healthchecks or probes.
+	ah = activatorhandler.NewMetricHandler(ctx, reporter, ah)
+	ah = activatorhandler.NewContextHandler(ctx, ah)
 
+	sigCtx, sigCancel := context.WithCancel(context.Background())
 	// Set up our health check based on the health of stat sink and environmental factors.
 	hc := newHealthCheck(sigCtx, logger, statSink)
 	ah = &activatorhandler.HealthHandler{HealthCheck: hc, NextHandler: ah, Logger: logger}
-
 	ah = network.NewProbeHandler(ah)
-
-	// NOTE: MetricHandler is being used as the outermost handler for the purpose of measuring the request latency.
-	ah = activatorhandler.NewMetricHandler(ctx, reporter, ah)
-	ah = activatorhandler.NewContextHandler(ctx, ah)
 
 	profilingHandler := profiling.NewHandler(logger, false)
 	// Watch the logging config map and dynamically update logging levels.