--- a/pkg/activator/handler/metric_handler_test.go
+++ b/pkg/activator/handler/metric_handler_test.go
@@ -17,6 +17,7 @@ package handler
 
 import (
 	"bytes"
+	"context"
 	"errors"
 	"net/http"
 	"net/http/httptest"
@@ -24,9 +25,9 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
+	"k8s.io/apimachinery/pkg/types"
 
 	rtesting "knative.dev/pkg/reconciler/testing"
-	"knative.dev/serving/pkg/activator"
 	"knative.dev/serving/pkg/network"
 )
 
@@ -101,13 +102,10 @@ func TestRequestMetricHandler(t *testing.T) {
 				cancel()
 			}()
 			reporter := &fakeReporter{}
-			revisionInformer(ctx, revision(testNamespace, testRevName))
 			handler := NewMetricHandler(ctx, reporter, test.baseHandler)
 
 			resp := httptest.NewRecorder()
 			req := httptest.NewRequest(http.MethodPost, "http://example.com", bytes.NewBufferString(""))
-			req.Header.Add(activator.RevisionHeaderNamespace, testNamespace)
-			req.Header.Add(activator.RevisionHeaderName, testRevName)
 			if test.newHeader != nil && len(test.newHeader) != 0 {
 				for k, v := range test.newHeader {
 					req.Header.Add(k, v)
@@ -128,7 +126,9 @@ func TestRequestMetricHandler(t *testing.T) {
 				}
 			}()
 
-			handler.ServeHTTP(resp, req)
+			reqCtx := withRevision(context.Background(), revision(testNamespace, testRevName))
+			reqCtx = withRevID(reqCtx, types.NamespacedName{Namespace: testNamespace, Name: testRevName})
+			handler.ServeHTTP(resp, req.WithContext(reqCtx))
 		})
 	}