<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

## Proposed Changes

Many of our inner activator handlers do the same things over and over again (fetch revision, create a keyed logger etc.). This avoids quite some extra work that each of these handlers do on their own in favor of doing it once in a "context handler" and adding all the reusable data onto the request's context, which is passed down automatically.

This spares us 13 of 60 allocations we do just in the innermost handler currently.

### Old

```
goos: darwin
goarch: amd64
pkg: knative.dev/serving/pkg/activator/handler
BenchmarkHandler/002k-resp-len-12         	  100771	     11956 ns/op	    9513 B/op	      60 allocs/op
BenchmarkHandler/016k-resp-len-12         	   69216	     18579 ns/op	   38484 B/op	      60 allocs/op
BenchmarkHandler/032k-resp-len-12         	   52394	     26079 ns/op	   71210 B/op	      60 allocs/op
BenchmarkHandler/064k-resp-len-12         	   23259	     45945 ns/op	  201494 B/op	      61 allocs/op
BenchmarkHandler/128k-resp-len-12         	    9903	    101044 ns/op	  496514 B/op	      62 allocs/op
PASS
```

### New

```
goos: darwin
goarch: amd64
pkg: knative.dev/serving/pkg/activator/handler
BenchmarkHandler/002k-resp-len-12         	  119955	     10514 ns/op	    7869 B/op	      47 allocs/op
BenchmarkHandler/016k-resp-len-12         	   66926	     17647 ns/op	   36360 B/op	      47 allocs/op
BenchmarkHandler/032k-resp-len-12         	   51644	     21144 ns/op	   69509 B/op	      47 allocs/op
BenchmarkHandler/064k-resp-len-12         	   26800	     49842 ns/op	  199945 B/op	      48 allocs/op
BenchmarkHandler/128k-resp-len-12         	   12469	     96675 ns/op	  498649 B/op	      49 allocs/op
PASS
```

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```